test(fix test): fix test for php8 compatility
