+++ tests/test-api-auth-upgrade-plan.php
@@ -13,7 +13,7 @@ class TestApiAuthUpgradePlan extends WP_UnitTestCase {
 	/**
 	 * Setup function.
 	 */
-	public function SetUp(){
+	public function SetUp(): void {
 		parent::setUp();
 		$this->instance = new ApiAuthUpgradePlan;
 	}
