+++ tests/test-gutenberg-transcription-block.php
@@ -13,7 +13,7 @@ class TestTranscriptionBlock extends WP_UnitTestCase {
     /**
      * Setup function.
      */
-    public function SetUp(){
+    public function SetUp(): void {
         parent::setUp();
         $this->instance = new TranscriptionBlock;
     }
