+++ tests/test-helpers.php
@@ -13,7 +13,7 @@ class TestHelpers extends WP_UnitTestCase {
     /**
      * Setup function.
      */
-    public function SetUp(){
+    public function SetUp(): void {
         parent::setUp();
         $this->instance = new Helpers;
     }
