+++ tests/test-listing-subtitle.php
@@ -13,7 +13,7 @@ class TestListingSubtitles extends WP_UnitTestCase {
     /**
      * Setup function.
      */
-    public function SetUp(){
+    public function SetUp(): void {
         parent::setUp();
         $this->instance = new ListingSubtitles;
     }
