+++ tests/test-media-library-extented.php
@@ -13,7 +13,7 @@ class TestMediaLibraryExtented extends WP_UnitTestCase {
     /**
      * Setup function.
      */
-    public function SetUp(){
+    public function SetUp(): void {
         parent::setUp();
         $this->instance = new MediaLibraryExtented;
     }
