+++ tests/test-transcriptions-extends.php
@@ -13,7 +13,7 @@ class TestTrascriptionsExtends extends WP_UnitTestCase {
     /**
      * Setup function.
      */
-    public function SetUp(){
+    public function SetUp(): void {
         parent::setUp();
         $this->instance = new TrascriptionsExtends;
     }
