+++ tests/test-video-block.php
@@ -13,7 +13,7 @@ class TestVideoBlock extends WP_UnitTestCase {
     /**
      * Setup function.
      */
-    public function SetUp(){
+    public function SetUp(): void {
         parent::setUp();
         $this->instance = new VideoBlock;
     }
